# internal/api/testdata/openapi.golden.json
{
  "components": {
    "schemas": {
...
      "CodeSigningKeylessRequest": {
        "properties": {
          "artifact_type": {
            "type": "string"
          },
          "digest": {
            "format": "byte",
            "type": "string"
          },
          "fulcio_issuer": {
            "type": "string"
          },
          "fulcio_san": {
            "type": "string"
          },
          "identity_method": {
            "type": "string"
          },
          "identity_payload": {
            "format": "byte",
            "type": "string"
          }
        },
        "required": [
          "artifact_type",
          "digest",
          "identity_method",
          "identity_payload"
        ],
        "type": "object"
      },
      "CodeSigningRequest": {
        "properties": {
          "artifact_type": {
            "type": "string"
          },
          "digest": {
            "format": "byte",
            "type": "string"
          },
          "key_id": {
            "type": "string"
          }
        },
        "required": [
          "key_id",
          "artifact_type",
          "digest"
        ],
        "type": "object"
      },
      "CodeSigningSignature": {
        "properties": {
          "algorithm": {
            "type": "string"
          },
          "artifact_type": {
            "type": "string"
          },
          "fulcio_issuer": {
            "type": "string"
          },
          "fulcio_san": {
            "type": "string"
          },
          "key_id": {
            "type": "string"
          },
          "public_key_der": {
            "format": "byte",
            "type": "string"
          },
          "signature": {
            "format": "byte",
            "type": "string"
          },
          "transparency_destination": {
            "type": "string"
          }
        },
        "required": [
          "algorithm",
          "artifact_type",
          "signature",
          "public_key_der"
        ],
        "type": "object"
      },
...
      "ManagedKey": {
        "properties": {
          "algorithm": {
            "type": "string"
          },
          "key_id": {
            "type": "string"
          },
          "public_der": {
            "format": "byte",
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "key_id",
          "algorithm",
          "version",
          "state"
        ],
        "type": "object"
      },
      "ManagedKeyActionRequest": {
        "properties": {
          "key_id": {
            "type": "string"
          }
        },
        "required": [
          "key_id"
        ],
        "type": "object"
      },
      "ManagedKeyGenerateRequest": {
        "properties": {
          "algorithm": {
            "type": "string"
          }
        },
        "required": [
          "algorithm"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "trstctl API token",
        "description": "Hashed API token resolved to a tenant-scoped principal with named trstctl permissions.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "description": "Verified browser session from OIDC, SAML, or LDAP; mutating requests also require the double-submit CSRF token.",
        "in": "cookie",
        "name": "trstctl_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Resource-oriented REST API for trstctl. Mutations require an Idempotency-Key; errors are RFC 7807 problem+json; lists use cursor pagination.",
    "title": "trstctl API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/code-signing/keyless": {
      "post": {
        "operationId": "signCodeArtifactKeyless",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeSigningKeylessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeSigningSignature"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Sign an artifact digest with a verified Sigstore/Fulcio identity",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/code-signing/sign": {
      "post": {
        "operationId": "signCodeArtifact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeSigningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeSigningSignature"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Sign an artifact digest with a managed code-signing key",
        "x-trstctl-permission": "keys:write"
      }
    },
...
    "/api/v1/managed-keys": {
      "post": {
        "operationId": "generateManagedKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedKeyGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedKey"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Generate a BYOK/HSM-resident managed key (private material stays in the provider)",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/managed-keys/revoke": {
      "post": {
        "operationId": "revokeManagedKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedKeyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedKey"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Revoke a managed key at the provider (requires dual-control approval)",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/managed-keys/rotate": {
      "post": {
        "operationId": "rotateManagedKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedKeyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedKey"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Rotate a managed key (mint a successor; requires dual-control approval)",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/managed-keys/zeroize": {
      "post": {
        "operationId": "zeroizeManagedKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedKeyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedKey"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Zeroize a managed key's material at the provider (requires dual-control approval)",
        "x-trstctl-permission": "keys:write"
      }
    },
...
  }
}